Create README.md
Synapse X Cracked
The complete, unredacted of the source code to a build of Synapse X dating from Juli 17, 2022.

Why have we decided to do this?
The source code fell into the wrong hands, and the one that perpetrated such a thing is willing to sell the source code for monetary gain. Rather than having some poor people waste hundreds of dollars for an older version of the source code to our project, we highly prefer releasing it for free and have it be an educational opportunity for many.
We are unsure if the perpetrator is actually willing to release the source code. However, it is best to take no chances and release it ourselves in his place to a guarantee a proper release.
The perpetrator that obtained the source code deals with malware on a daily basis. It would be insanely unsafe for people to download his version of the source code, because it could be packed with dangerous malware. Our own release is guaranteed to be free of viruses.
Our own version is properly documented. His version lacks the comments and additional information that explains the functioning of our code.
To protect our competitors. While it is expected of us to supplant our competitors, we do not condone unfair marketing practices or hurtful behavior towards them. The perpetrator has allegedly obtained identifying information about our competitors, and we do not want it to be included or otherwise spread by his releases.
What does the release contains?
The unredacted, complete source code of Synapse X dating from October 21, 2019. The code is obviously not updated, but should compile as is using Visual Studio 2019.
Our numerous libraries that aids us in developing Synapse X and associated products.
The source code to several scripts available in the Synapse X script hub, excluding proprietary (paid) scripts.
The complete source code to our user interfaces.
Licensing information
This source code is licensed under the AGPLv3. This means that if you release a project using some of this codebase, your release will have to contain its source code. If you do not include the source code, legal action can be taken. We welcome developers to study, share and improve this source code, but the license does not support its unauthorized usage in commercial projects and the likes.